fixed grunt and installation script

+++ js/apps/system/_admin/aardvark/APP/frontend/build/app.js
@@ -10901,7 +10901,9 @@ function GraphViewer(svg, width, height, adapterConfig, config) {
                 });
               }
               else {
-                this.deleteAllAardvarkJobs(); 
+                if (AaJobs.length > 0) {
+                  this.deleteAllAardvarkJobs(); 
+                }
               }
               callback(false, array);
             }
@@ -18581,11 +18583,12 @@ window.arangoDocument = Backbone.Collection.extend({
         query: query,
         bindVars: bindVars
       };
+      /*
       if (this.getTotal() < 10000 || this.filters.length > 0) {
         queryObj.options = {
           fullCount: true,
         };
-      }
+      }*/
 
       var checkCursorStatus = function(jobid) {
         $.ajax({
@@ -21746,7 +21749,7 @@ window.ArangoUsers = Backbone.Collection.extend({
     el2: '#collectionsThumbnailsIn',
 
     searchTimeout: null,
-    refreshRate: 2000,
+    refreshRate: 10000,
 
     template: templateEngine.createTemplate("collectionsView.ejs"),
 
@@ -21764,7 +21767,6 @@ window.ArangoUsers = Backbone.Collection.extend({
         var callback = function(error, lockedCollections) {
           var self = this;
           if (error) {
-            //arangoHelper.arangoError("Collections", "Could not check locked collections");
             console.log("Could not check locked collections");
           }
           else {
@@ -21791,6 +21793,8 @@ window.ArangoUsers = Backbone.Collection.extend({
               if (model.get("locked") || model.get("status") === 'loading') {
                 $('#collection_' + model.get("name")).addClass('locked');
                 if (model.get("locked")) {
+                  $('#collection_' + model.get("name")).find('.corneredBadge').removeClass('loaded unloaded');
+                  $('#collection_' + model.get("name")).find('.corneredBadge').addClass('inProgress');
                   $('#collection_' + model.get("name") + ' .corneredBadge').text(model.get("desc"));
                 }
                 else {
